Restructure app packages
