+++ .idea/misc.xml
@@ -6,7 +6,9 @@
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.16770833333333332" />
         <entry key="app/src/main/res/drawable/ic_day_theme.xml" value="0.16770833333333332" />
         <entry key="app/src/main/res/drawable/ic_night_theme.xml" value="0.16770833333333332" />
-        <entry key="app/src/main/res/layout/activity_add_note.xml" value="0.1393110435663627" />
+        <entry key="app/src/main/res/layout/activity_add_note.xml" value="0.25" />
+        <entry key="app/src/main/res/layout/activity_main.xml" value="0.1393110435663627" />
+        <entry key="app/src/main/res/layout/empty_state.xml" value="0.1393110435663627" />
         <entry key="app/src/main/res/layout/note_item.xml" value="0.1393110435663627" />
         <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.16770833333333332" />
       </map>
