+++ app/src/main/java/br/com/mmdevelopment/cleannotes/App.kt
@@ -1,8 +1,8 @@
 package br.com.mmdevelopment.cleannotes
 
 import android.app.Application
-import br.com.mmdevelopment.cleannotes.datasource.di.dataModule
-import br.com.mmdevelopment.cleannotes.ui.di.vmModule
+import br.com.mmdevelopment.cleannotes.di.dataModule
+import br.com.mmdevelopment.cleannotes.di.vmModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
